+++ examples/mario.js
@@ -1,31 +1,22 @@
 kaboom({
-	scale: 2,
 	debug: true,
 	clearColor: [0, 0, 0, 1],
 });
 
 // load assets
-loadRoot("/pub/examples/");
-loadSprite("guy", "img/guy.png");
-loadSprite("spike", "img/spike.png");
-loadSprite("steel", "img/steel.png");
-loadSprite("prize", "img/jumpy.png");
-loadSprite("apple", "img/apple.png");
-loadSprite("coin", "img/coin.png");
+loadSprite("bean", "sprites/bean.png");
+loadSprite("spike", "sprites/spike.png");
+loadSprite("steel", "sprites/steel.png");
+loadSprite("prize", "sprites/jumpy.png");
+loadSprite("apple", "sprites/apple.png");
+loadSprite("coin", "sprites/coin.png");
 
 // define some constants
-const JUMP_FORCE = 320;
-const MOVE_SPEED = 120;
-const FALL_DEATH = 640;
+const JUMP_FORCE = 1440;
+const MOVE_SPEED = 480;
+const FALL_DEATH = 2400;
 
-// define layers, draw "ui" on top, and "game" is the default layer
-layers([
-	"game",
-	"ui",
-], "game");
-
-// camera will ignore "ui" layer
-camIgnore([ "ui", ]);
+gravity(3200);
 
 // add level to scene
 const level = addLevel([
@@ -38,52 +29,50 @@ const level = addLevel([
 ], {
 	// TODO: derive grid size from sprite size instead of hardcode
 	// grid size
-	width: 11,
-	height: 11,
+	width: 64,
+	height: 64,
 	// define each object as a list of components
 	"=": [
 		sprite("steel"),
 		area(),
 		solid(),
+		origin("bot"),
 	],
 	"$": [
 		sprite("coin"),
 		area(),
+		origin("bot"),
 		"coin",
 	],
 	"%": [
 		sprite("prize"),
 		area(),
 		solid(),
+		origin("bot"),
 		"prize",
 	],
 	"^": [
 		sprite("spike"),
 		area(vec2(0, 6), vec2(11, 11)),
 		area(),
+		body(),
+		origin("bot"),
 		"dangerous",
 	],
 	"#": [
 		sprite("apple"),
 		area(),
+		origin("bot"),
+		body(),
 		"apple",
 	],
 });
-
-// add score counter obj
-const score = add([
-	text("0"),
-	pos(6, 6),
-	layer("ui"),
-	{
-		value: 0,
-	},
-]);
 
 // define a custom component that handles player grow big logic
 function big() {
 	let timer = 0;
 	let isBig = false;
+	let destScale = 1;
 	return {
 		// component id / name
 		id: "big",
@@ -97,18 +86,19 @@ function big() {
 					this.smallify();
 				}
 			}
+			this.scale = this.scale.lerp(vec2(destScale), dt() * 6);
 		},
 		// custom methods
 		isBig() {
 			return isBig;
 		},
 		smallify() {
-			this.scale = vec2(1);
+			destScale = 1;
 			timer = 0;
 			isBig = false;
 		},
 		biggify(time) {
-			this.scale = vec2(2);
+			destScale = 2;
 			timer = time;
 			isBig = true;
 		},
@@ -117,7 +107,7 @@ function big() {
 
 // define player object
 const player = add([
-	sprite("guy"),
+	sprite("bean"),
 	pos(0, 0),
 	area(),
 	scale(1),
@@ -125,6 +115,7 @@ const player = add([
 	body(),
 	// the custom component we defined above
 	big(),
+	origin("bot"),
 ]);
 
 // action() runs every frame
@@ -133,19 +124,20 @@ player.action(() => {
 	camPos(player.pos);
 	// check fall death
 	if (player.pos.y >= FALL_DEATH) {
-		go("lose", { score: score.value, });
+		go("lose");
 	}
 });
 
 // if player collides with any obj with "dangerous" tag, lose
 player.collides("dangerous", () => {
-	go("lose", { score: score.value, });
+	go("lose");
 });
 
 // grow an apple if player's head bumps into an obj with "prize" tag
 player.on("headbutt", (obj) => {
 	if (obj.is("prize")) {
-		level.spawn("#", obj.gridPos.sub(0, 1));
+		const apple = level.spawn("#", obj.gridPos.sub(0, 1));
+		apple.jump();
 	}
 });
 
@@ -156,11 +148,8 @@ player.collides("apple", (a) => {
 	player.biggify(3);
 });
 
-// increase score if meets coin
 player.collides("coin", (c) => {
 	destroy(c);
-	score.value++;
-	score.text = score.value;
 });
 
 // jump with space
@@ -179,10 +168,8 @@ keyDown("right", () => {
 	player.move(MOVE_SPEED, 0);
 });
 
-scene("lose", ({ score }) => {
+scene("lose", () => {
 	add([
-		text("Score: " + score, 24),
-		origin("center"),
-		pos(width() / 2, height() / 2),
+		text("you lose"),
 	]);
 });