+++ examples/multiboom.js
@@ -5,44 +5,39 @@ const clearColors = [
 	[0, 0, 1, 1],
 ];
 
-const sprites = [
-	"mark",
-	"notmark",
-];
-
 for (let i = 0; i < 2; i++) {
 
 	const k = kaboom({
 		clearColor: clearColors[i],
+		noGlobal: true,
 		width: 320,
 		height: 320,
 	});
 
-	k.loadRoot("/pub/examples/");
-	k.loadSprite(sprites[i], `img/${sprites[i]}.png`);
+	k.loadBean();
 
 	// custom spin component
 	function spin() {
 		return {
 			id: "spin",
 			update() {
-				this.scale = Math.sin(k.time()) * 24;
+				this.scale = Math.sin(k.time() + i) * 12;
 				this.angle = k.time();
 			},
 		};
 	}
 
 	k.add([
-		k.sprite(sprites[i]),
+		k.sprite("bean"),
 		k.pos(k.width() / 2, k.height() / 2),
-		k.scale(12),
+		k.scale(6),
 		k.rotate(0),
 		spin(),
 		k.origin("center"),
 	]);
 
 	k.add([
-		k.text(`kaboom #${i}`, 24),
+		k.text(`#${i}`),
 	]);
 
 }